requirements.txt
flask
flask_sqlalchemy
cryptography
opencv-python
pillow
pyjwt
werkzeug
